requirements.txt
fastapi
uvicorn
google-generativeai
requests